% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_network.R
\name{build_mi_graph}
\alias{build_mi_graph}
\title{Build MI Graph}
\usage{
build_mi_graph(
  mi_matrix,
  threshold_method = "percentile",
  percentile = 0.95,
  fixed_threshold = NULL
)
}
\arguments{
\item{mi_matrix}{Mutual information matrix}

\item{threshold_method}{Threshold method, either "percentile" or "fixed"}

\item{percentile}{Percentile for thresholding (used if threshold_method = "percentile")}

\item{fixed_threshold}{Fixed MI value for thresholding (used if threshold_method = "fixed")}
}
\value{
igraph object
}
\description{
Build MI Graph
}
