% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose_adaptive_nbins}
\alias{choose_adaptive_nbins}
\title{Choose Adaptive Number of Bins}
\usage{
choose_adaptive_nbins(
  expr_data,
  min_bins = 5,
  max_bins = 30,
  min_cells_per_bin = 75
)
}
\arguments{
\item{expr_data}{Expression data}

\item{min_bins}{Minimum number of bins}

\item{max_bins}{Maximum number of bins}

\item{min_cells_per_bin}{Minimum cells per bin}
}
\value{
Number of bins
}
\description{
Choose Adaptive Number of Bins
}
