% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{compute_cmi_matrix_parallel}
\alias{compute_cmi_matrix_parallel}
\title{Compute Conditional MI Matrix in Parallel}
\usage{
compute_cmi_matrix_parallel(
  expr_subset,
  pt_vec,
  n_cores = parallel::detectCores() - 10,
  nbins = 10
)
}
\arguments{
\item{expr_subset}{Expression matrix (genes x cells)}

\item{pt_vec}{Pseudotime vector (length = cells)}

\item{n_cores}{Cores}

\item{nbins}{Discretization bins}
}
\value{
Conditional MI matrix
}
\description{
Compute Conditional MI Matrix in Parallel
}
