% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{degree_num}
\alias{degree_num}
\title{Degree as Numeric Vector}
\usage{
degree_num(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
Numeric vector of degrees.
}
\description{
Degree as Numeric Vector
}
\keyword{internal}

