% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{edge_gain_loss}
\alias{edge_gain_loss}
\title{Edge Gain/Loss Analysis}
\usage{
edge_gain_loss(bin_graphs)
}
\arguments{
\item{bin_graphs}{List of graphs per bin (NULL allowed)}
}
\value{
List of stable, gained, lost edge keys
}
\description{
Edge Gain/Loss Analysis
}
