% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_housekeeping}
\alias{filter_housekeeping}
\title{Filter Housekeeping Genes}
\usage{
filter_housekeeping(gene_list)
}
\arguments{
\item{gene_list}{Character vector of gene names.}
}
\value{
Character vector of filtered gene names.
}
\description{
Removes common housekeeping / technical genes (ribosomal, GAPDH/ACTB, mitochondrial).
}
