% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{gene_pt_correlation}
\alias{gene_pt_correlation}
\title{Gene-Pseudotime Correlation}
\usage{
gene_pt_correlation(
  seurat_obj,
  pseudotime_col,
  assay = "RNA",
  layer = "data",
  method = "spearman"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{pseudotime_col}{Metadata column name for pseudotime}

\item{assay}{Assay name}

\item{layer}{Layer (Seurat v5) or slot (Seurat v4)}

\item{method}{Correlation method}
}
\value{
Data frame with gene correlations
}
\description{
Gene-Pseudotime Correlation
}
