% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergraph.R
\name{make_hypergraph}
\alias{make_hypergraph}
\title{Create Hypergraph from Gene Groups}
\usage{
make_hypergraph(group_list, all_genes)
}
\arguments{
\item{group_list}{List of gene groups}

\item{all_genes}{All genes}
}
\value{
List containing hypergraph, incidence matrix, genes, and groups
}
\description{
Create Hypergraph from Gene Groups
}
