% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{make_pt_bins}
\alias{make_pt_bins}
\title{Bin Cells Along Pseudotime}
\usage{
make_pt_bins(seurat_obj, pseudotime_col, n_bins = 5, bin_method = "quantile")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{pseudotime_col}{Metadata column name for pseudotime}

\item{n_bins}{Number of bins}

\item{bin_method}{Binning method ("quantile" or "equal_width")}
}
\value{
Vector of bin assignments
}
\description{
Bin Cells Along Pseudotime
}
