% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_terms.R
\name{plot_go_terms}
\alias{plot_go_terms}
\title{Plot Enrichment Terms}
\usage{
plot_go_terms(
  enrich_results,
  top_n_terms = 4,
  title = "GO Terms for Communities",
  ontology = "BP"
)
}
\arguments{
\item{enrich_results}{A named list of enrichment results from enrich_go.}

\item{top_n_terms}{Integer, number of top terms to plot per community (default: 4).}

\item{title}{Character, title of the plot (default: "GO Terms for Communities").}

\item{ontology}{Character, ontology used ("BP", "CC", "MF", "ALL", "KEGG", "DO", "Reactome") (default: "BP").}
}
\value{
A ggplot object.
}
\description{
Visualizes the top enrichment terms from enrichment results in a dot plot.
}
