% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hypergraph.R
\name{plot_hypergraph}
\alias{plot_hypergraph}
\title{Plot Hypergraph with Overlapping Genes Highlighted}
\usage{
plot_hypergraph(
  large_communities,
  comm_indices = NULL,
  title = "Hypergraph with Overlapping Communities",
  key_genes = c("PRSS1", "PRSS2")
)
}
\arguments{
\item{large_communities}{A list of communities, where each element contains the nodes of a community.}

\item{comm_indices}{Integer vector, indices of communities to plot (default: all communities).}

\item{title}{Character, title of the plot (default: "Hypergraph with Overlapping Communities").}

\item{key_genes}{Character vector, additional genes to label (default: c("PRSS1", "PRSS2")).}
}
\value{
A ggraph plot object.
}
\description{
Visualizes a hypergraph where communities are hyperedges, highlighting overlapping genes between specified communities.
}
