% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mi_network.R
\name{plot_mi_network}
\alias{plot_mi_network}
\title{Plot MI Network with Community Coloring}
\usage{
plot_mi_network(
  g_consensus,
  large_communities,
  sample_size = 100,
  top_n_per_comm = 5,
  title = "Network Communities"
)
}
\arguments{
\item{g_consensus}{An igraph object representing the consensus graph.}

\item{large_communities}{A list of communities, where each element contains the nodes of a community.}

\item{sample_size}{Integer, number of nodes to sample from the graph (default: 100).}

\item{top_n_per_comm}{Integer, number of top nodes per community to label (default: 5).}

\item{title}{Character, title of the plot (default: "Network Communities").}
}
\value{
A ggraph plot object.
}
\description{
Visualizes a subset of an MI network with community coloring, labeling top nodes from each community.
}
