% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_network.R
\name{compute_mi_matrix_parallel}
\alias{compute_mi_matrix_parallel}
\title{Compute Mutual Information Matrix in Parallel}
\usage{
compute_mi_matrix_parallel(expr_subset, n_cores = parallel::detectCores() - 10)
}
\arguments{
\item{expr_subset}{Expression data subset}

\item{n_cores}{Number of cores}
}
\value{
MI matrix
}
\description{
Compute Mutual Information Matrix in Parallel
}
