% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_housekeeping}
\alias{filter_housekeeping}
\title{Filter Housekeeping Genes}
\usage{
filter_housekeeping(gene_list)
}
\arguments{
\item{gene_list}{Vector of gene names}
}
\value{
Filtered gene list excluding housekeeping genes
}
\description{
Filter Housekeeping Genes
}
